requirements.txt
asgiref==3.4.1
certifi==2025.1.31
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.0
drf-yasg==1.20.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
psycopg2==2.9.8
pyparsing==3.1.4
pytz==2025.1
requests==2.27.1
ruamel.yaml==0.18.3
ruamel.yaml.clib==0.2.8
sqlparse==0.4.4
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.20
